refactor: renamed the i18n functions
