+++ packages/dantalion-i18n/src/resources/getAsync.ts
